content: remove content component
its responsabillity should be on the checkout component

it also fixes parameters on the checkout component that were hardcoded
