+++ webapp/src/components/Checkout/style.css
@@ -24,3 +24,8 @@
   animation-duration: 450ms;
   animation-timing-function: cubic-bezier(0.23, 1, 0.32, 1);
 }
+
+.content {
+  flex: 1;
+}
+
