--- webapp/src/components/Content/styles.css
@@ -1,5 +1,0 @@
-@import "../../styles/spacing.css";
-
-.content {
-  flex: 1;
-}
