--- webapp/stories/Content/styles.css
@@ -1,11 +1,0 @@
-.page {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 200px;
-  margin-bottom: 30px;
-  color: #ffffff;
-  text-align: center;
-  background-color: #333333;
-}
