+++ webapp/stories/index.js
@@ -4,6 +4,5 @@ import './Grid'
 import './Header'
 import './Typography'
 import './FormFields'
-import './Content'
 import './ProgressBar'
 import './Footer'
